# sql: reduce indexed variable allocations