--- a/pkg/sql/catalog/schemaexpr/computed_exprs.go
+++ b/pkg/sql/catalog/schemaexpr/computed_exprs.go
@@ -11,8 +11,6 @@
 package schemaexpr
 
 import (
-	"context"
-
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog"
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgcode"
 	"github.com/cockroachdb/cockroach/pkg/sql/pgwire/pgerror"
@@ -35,9 +33,7 @@ type RowIndexedVarContainer struct {
 var _ eval.IndexedVarContainer = &RowIndexedVarContainer{}
 
 // IndexedVarEval implements eval.IndexedVarContainer.
-func (r *RowIndexedVarContainer) IndexedVarEval(
-	ctx context.Context, idx int, e tree.ExprEvaluator,
-) (tree.Datum, error) {
+func (r *RowIndexedVarContainer) IndexedVarEval(idx int) (tree.Datum, error) {
 	rowIdx, ok := r.Mapping.Get(r.Cols[idx].GetID())
 	if !ok {
 		return tree.DNull, nil