--- a/pkg/sql/colexec/colexectestutils/proj_utils.go
+++ b/pkg/sql/colexec/colexectestutils/proj_utils.go
@@ -33,10 +33,8 @@ type MockTypeContext struct {
 var _ eval.IndexedVarContainer = &MockTypeContext{}
 
 // IndexedVarEval implements the eval.IndexedVarContainer interface.
-func (p *MockTypeContext) IndexedVarEval(
-	ctx context.Context, idx int, e tree.ExprEvaluator,
-) (tree.Datum, error) {
-	return tree.DNull.Eval(ctx, e)
+func (p *MockTypeContext) IndexedVarEval(idx int) (tree.Datum, error) {
+	return tree.DNull, nil
 }
 
 // IndexedVarResolvedType implements the tree.IndexedVarContainer interface.