--- a/pkg/sql/distsql_plan_stats.go
+++ b/pkg/sql/distsql_plan_stats.go
@@ -504,9 +504,6 @@ func (dsp *DistSQLPlanner) createStatsPlan(
 
 		var rb renderBuilder
 		rb.init(exec.Node(planNode(&scan)), exec.OutputOrdering{})
-		for i, expr := range exprs {
-			exprs[i] = rb.r.ivarHelper.RebindTyped(expr)
-		}
 		rb.setOutput(exprs, resultCols)
 
 		err = dsp.createPlanForRender(ctx, p, rb.r, planCtx)