--- a/pkg/sql/execinfrapb/expr_test.go
+++ b/pkg/sql/execinfrapb/expr_test.go
@@ -22,25 +22,21 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/util/leaktest"
 )
 
-type testVarContainer struct{}
-
-var _ tree.IndexedVarContainer = &testVarContainer{}
-
-func (d *testVarContainer) IndexedVarResolvedType(idx int) *types.T {
-	return types.Int
-}
-
-func TestProcessExpression(t *testing.T) {
+func TestDeserializeExpr(t *testing.T) {
 	defer leaktest.AfterTest(t)()
 
 	e := Expression{Expr: "@1 * (@2 + @3) + @1"}
 
-	var c testVarContainer
-	h := tree.MakeIndexedVarHelper(&c, 4)
 	st := cluster.MakeTestingClusterSettings()
 	evalCtx := eval.MakeTestingEvalContext(st)
 	semaCtx := tree.MakeSemaContext()
-	expr, err := processExpression(context.Background(), e, &evalCtx, &semaCtx, &h)
+	expr, err := DeserializeExpr(
+		context.Background(),
+		e,
+		[]*types.T{types.Int, types.Int, types.Int},
+		&semaCtx,
+		&evalCtx,
+	)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -59,7 +55,13 @@ func TestProcessExpression(t *testing.T) {
 
 	// We can process a new expression with the same tree.IndexedVarHelper.
 	e = Expression{Expr: "@4 - @1"}
-	expr, err = processExpression(context.Background(), e, &evalCtx, &semaCtx, &h)
+	expr, err = DeserializeExpr(
+		context.Background(),
+		e,
+		[]*types.T{types.Int, types.Int, types.Int, types.Int},
+		&semaCtx,
+		&evalCtx,
+	)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -79,16 +81,21 @@ func TestProcessExpression(t *testing.T) {
 }
 
 // Test that processExpression evaluates constant exprs into datums.
-func TestProcessExpressionConstantEval(t *testing.T) {
+func TestDeserializeExpressionConstantEval(t *testing.T) {
 	defer leaktest.AfterTest(t)()
 
 	e := Expression{Expr: "ARRAY[1:::INT,2:::INT]"}
 
-	h := tree.MakeIndexedVarHelper(nil, 0)
 	st := cluster.MakeTestingClusterSettings()
 	evalCtx := eval.MakeTestingEvalContext(st)
 	semaCtx := tree.MakeSemaContext()
-	expr, err := processExpression(context.Background(), e, &evalCtx, &semaCtx, &h)
+	expr, err := DeserializeExpr(
+		context.Background(),
+		e,
+		[]*types.T{types.Int, types.Int},
+		&semaCtx,
+		&evalCtx,
+	)
 	if err != nil {
 		t.Fatal(err)
 	}