--- a/pkg/sql/sem/eval/indexed_vars.go
+++ b/pkg/sql/sem/eval/indexed_vars.go
@@ -10,16 +10,12 @@
 
 package eval
 
-import (
-	"context"
-
-	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
-)
+import "github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 
 // IndexedVarContainer extends tree.IndexedVarContainer with the ability to
 // evaluate the underlying expression.
 type IndexedVarContainer interface {
 	tree.IndexedVarContainer
 
-	IndexedVarEval(ctx context.Context, idx int, e tree.ExprEvaluator) (tree.Datum, error)
+	IndexedVarEval(idx int) (tree.Datum, error)
 }