--- a/pkg/sql/sem/transform/BUILD.bazel
+++ b/pkg/sql/sem/transform/BUILD.bazel
@@ -2,16 +2,12 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
     name = "transform",
-    srcs = [
-        "aggregates.go",
-        "expr_transform.go",
-    ],
+    srcs = ["expr_transform.go"],
     importpath = "github.com/cockroachdb/cockroach/pkg/sql/sem/transform",
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/sql/sem/eval",
         "//pkg/sql/sem/normalize",
         "//pkg/sql/sem/tree",
-        "//pkg/sql/sessiondata",
     ],
 )