--- a/pkg/sql/sem/transform/expr_transform.go
+++ b/pkg/sql/sem/transform/expr_transform.go
@@ -16,7 +16,6 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/eval"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/normalize"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
-	"github.com/cockroachdb/cockroach/pkg/sql/sessiondata"
 )
 
 // ExprTransformContext supports the methods that test expression
@@ -24,8 +23,7 @@ import (
 // should be used in planner instance to avoid re-allocation of these
 // visitors between uses.
 type ExprTransformContext struct {
-	normalizeVisitor   normalize.Visitor
-	isAggregateVisitor isAggregateVisitor
+	normalizeVisitor normalize.Visitor
 }
 
 // NormalizeExpr is a wrapper around EvalContex.Expr which
@@ -43,24 +41,4 @@ func (t *ExprTransformContext) NormalizeExpr(
 		return nil, err
 	}
 	return expr.(tree.TypedExpr), nil
-}
-
-// AggregateInExpr determines if an Expr contains an aggregate function.
-// TODO(knz/radu): this is not the right way to go about checking
-// these things. Instead whatever analysis occurs prior on the expression
-// should collect scalar properties (see tree.ScalarProperties) and
-// then the collected properties should be tested directly.
-func (t *ExprTransformContext) AggregateInExpr(
-	ctx context.Context, expr tree.Expr, searchPath sessiondata.SearchPath,
-) bool {
-	if expr == nil {
-		return false
-	}
-
-	t.isAggregateVisitor = isAggregateVisitor{
-		ctx:        ctx,
-		searchPath: searchPath,
-	}
-	tree.WalkExprConst(&t.isAggregateVisitor, expr)
-	return t.isAggregateVisitor.Aggregated
 }