… pkg/sql/sem/transform/aggregates.go
… pkg/sql/sem/tree/format_test.go
… pkg/sql/sem/tree/indexed_vars_test.go