#### execinfrapb: remove unnecessary DefaultIntSize parse option

When `execinfrapb.Expression`s are serialized, references to integer
types are always formatted unambiguously with the
`tree.FmtCheckEquivalence` flag as `INT2`, `INT4`, or `INT8`, and never
as `INT`. Therefore, there is never a need to disambiguate the meaning
of `INT` based on the `default_int_size` session setting. This allows
simplification of the deserialization logic.

Release note: None

#### execinfrapb: remove unnecessary check for aggregates

This commit removes some logic in `execinfrapb` that asserted that
expressions being deserialized do not contain aggregate functions. This
logics appears to be vestigial that is no longer required, though I was
unable to pinpoint the original motivation behind it. This type of
semantic analysis should now be performed during optimization-time, so I
see no reason for it to remain.

This logic was the last remaining user of the `isAggregateVisitor` so it
has been removed.

Release note: None

#### execinfrapb: do not use tree.IndexedVarHelper

Previously, a `tree.IndexedVarHelper` was used to "rebind" indexed
variables in expressions deserialized in `execinfrapb`. The only real
function of this rebinding was to assign types to indexed variables so
that they could be type-checked. However, this was an unnecessary step
because the type checker can already determine the types of indexed vars
via the `IVarContainer` of the `tree.SemaContext`. The `exprHelper` type
already implements `tree.IndexedVarContainer` so it was already possible
to simply type-check indexed variables without rebinding them first.

This reduces allocations of `tree.IndexedVarHelper`s that can be large,
especially when there are many indexed variables.

Release note: None

#### sql/sem/tree: remove IndexedVarHelper.RebindTyped

Indexed variable rebinding is no longer needed, so it has been removed.
This reduces unnecessary allocations because `tree.IndexedVarHelper`s no
longer need to be created to perform the rebinding.

Informs #117546

Release note: None

#### sql: remove extra params for IndexedVarContainer.IndexedVarEval

This method signature for `IndexedVarContainer.IndexedVarEval` by
removing unneeded parameters. It also removes some unnecessary calls to
`Eval` on `tree.Datum`s.

Release note: None

#### sql: remove unnecessary invocation of makePredicate

Release note: None