@mgartner 2024-02-23
The first two commits (currently) are from #119557. 

The flow setup benchmarks show some improvements:

```
name                                                 old time/op    new time/op    delta
FlowSetup/vectorize=true/distribute=true-10            51.5µs ± 0%    51.3µs ± 1%     ~     (p=0.222 n=5+5)
FlowSetup/vectorize=true/distribute=false-10           49.8µs ± 0%    49.8µs ± 1%     ~     (p=0.690 n=5+5)
FlowSetup/vectorize=false/distribute=true-10           49.7µs ± 0%    49.6µs ± 0%     ~     (p=0.310 n=5+5)
FlowSetup/vectorize=false/distribute=false-10          50.5µs ±12%    50.0µs ±12%     ~     (p=0.310 n=5+5)
FlowRenderSetup/vectorize=true/distribute=true-10      1.54ms ± 1%    1.52ms ± 0%   -1.71%  (p=0.008 n=5+5)
FlowRenderSetup/vectorize=true/distribute=false-10     1.38ms ± 0%    1.35ms ± 0%   -1.94%  (p=0.008 n=5+5)
FlowRenderSetup/vectorize=false/distribute=true-10     1.37ms ± 0%    1.32ms ± 0%   -3.54%  (p=0.016 n=5+4)
FlowRenderSetup/vectorize=false/distribute=false-10    1.21ms ± 0%    1.17ms ± 0%   -3.96%  (p=0.008 n=5+5)

name                                                 old alloc/op   new alloc/op   delta
FlowSetup/vectorize=true/distribute=true-10            23.1kB ± 2%    23.0kB ± 2%     ~     (p=0.548 n=5+5)
FlowSetup/vectorize=true/distribute=false-10           21.6kB ± 2%    21.4kB ± 3%     ~     (p=0.548 n=5+5)
FlowSetup/vectorize=false/distribute=true-10           28.1kB ± 2%    27.4kB ± 2%     ~     (p=0.222 n=5+5)
FlowSetup/vectorize=false/distribute=false-10          26.8kB ± 3%    26.2kB ± 3%     ~     (p=0.486 n=4+4)
FlowRenderSetup/vectorize=true/distribute=true-10       982kB ± 2%     942kB ± 2%   -4.07%  (p=0.008 n=5+5)
FlowRenderSetup/vectorize=true/distribute=false-10      897kB ± 0%     862kB ± 2%   -3.96%  (p=0.016 n=4+5)
FlowRenderSetup/vectorize=false/distribute=true-10      747kB ± 0%     667kB ± 0%  -10.82%  (p=0.029 n=4+4)
FlowRenderSetup/vectorize=false/distribute=false-10     671kB ± 0%     591kB ± 0%  -12.02%  (p=0.008 n=5+5)

name                                                 old allocs/op  new allocs/op  delta
FlowSetup/vectorize=true/distribute=true-10               236 ± 1%       234 ± 1%     ~     (p=0.222 n=5+5)
FlowSetup/vectorize=true/distribute=false-10              218 ± 2%       217 ± 1%     ~     (p=0.333 n=5+5)
FlowSetup/vectorize=false/distribute=true-10              235 ± 1%       231 ± 1%     ~     (p=0.079 n=5+5)
FlowSetup/vectorize=false/distribute=false-10             219 ± 2%       215 ± 2%     ~     (p=0.114 n=4+4)
FlowRenderSetup/vectorize=true/distribute=true-10       9.19k ± 3%     8.69k ± 4%   -5.46%  (p=0.008 n=5+5)
FlowRenderSetup/vectorize=true/distribute=false-10      5.81k ± 0%     5.34k ± 2%   -8.02%  (p=0.016 n=4+5)
FlowRenderSetup/vectorize=false/distribute=true-10      5.55k ± 0%     4.54k ± 0%  -18.09%  (p=0.029 n=4+4)
FlowRenderSetup/vectorize=false/distribute=false-10     2.28k ± 0%     1.28k ± 0%  -43.90%  (p=0.008 n=5+5)
```

And there's some improvement in the end-to-end benchmarks:

```
name                                              old time/op    new time/op    delta
EndToEnd/kv-read/Simple-10                           143µs ± 1%     143µs ± 2%     ~     (p=1.000 n=5+5)
EndToEnd/kv-read/Prepared-10                        86.9µs ± 3%    86.9µs ± 3%     ~     (p=0.690 n=5+5)
EndToEnd/kv-read-const/Simple-10                    98.7µs ± 1%    99.7µs ± 2%     ~     (p=0.095 n=5+5)
EndToEnd/kv-read-const/Prepared-10                  91.9µs ± 1%    92.2µs ± 3%     ~     (p=0.841 n=5+5)
EndToEnd/tpcc-new-order/Simple-10                    183µs ± 0%     184µs ± 0%   +0.41%  (p=0.016 n=5+5)
EndToEnd/tpcc-new-order/Prepared-10                 98.7µs ± 2%    98.9µs ± 3%     ~     (p=0.690 n=5+5)
EndToEnd/tpcc-delivery/Simple-10                     258µs ±12%     244µs ± 0%     ~     (p=0.730 n=5+4)
EndToEnd/tpcc-delivery/Prepared-10                   154µs ± 1%     156µs ± 1%     ~     (p=0.190 n=5+4)
EndToEnd/tpcc-stock-level/Simple-10                  482µs ± 1%     483µs ± 0%     ~     (p=0.686 n=4+4)
EndToEnd/tpcc-stock-level/Prepared-10                301µs ± 1%     301µs ± 1%     ~     (p=0.310 n=5+5)
EndToEnd/many-columns-and-indexes-a/Simple-10        672µs ± 0%     673µs ± 0%     ~     (p=0.413 n=5+4)
EndToEnd/many-columns-and-indexes-a/Prepared-10      565µs ± 2%     563µs ± 1%     ~     (p=0.841 n=5+5)
EndToEnd/many-columns-and-indexes-b/Simple-10        705µs ± 1%     704µs ± 0%     ~     (p=0.841 n=5+5)
EndToEnd/many-columns-and-indexes-b/Prepared-10      582µs ± 1%     581µs ± 1%     ~     (p=0.841 n=5+5)
EndToEnd/many-columns-and-indexes-c/Simple-10        190ms ± 1%     177ms ± 1%   -6.70%  (p=0.029 n=4+4)
EndToEnd/many-columns-and-indexes-c/Prepared-10      179ms ± 6%     176ms ± 0%     ~     (p=0.690 n=5+5)
EndToEnd/single-col-histogram-range/Simple-10        231µs ± 1%     231µs ± 0%     ~     (p=0.690 n=5+5)
EndToEnd/single-col-histogram-range/Prepared-10      129µs ± 1%     130µs ± 2%     ~     (p=0.548 n=5+5)
EndToEnd/batch-insert-one/Simple-10                 29.3ms ± 1%    29.3ms ± 2%     ~     (p=0.548 n=5+5)
EndToEnd/batch-insert-one/Prepared-10               29.5ms ± 0%    29.6ms ± 2%     ~     (p=0.548 n=5+5)
EndToEnd/batch-insert-many/Simple-10                35.0ms ± 1%    35.2ms ± 1%     ~     (p=0.841 n=5+5)
EndToEnd/batch-insert-many/Prepared-10              35.8ms ±14%    35.9ms ±14%     ~     (p=1.000 n=5+5)
EndToEnd/ored-preds-100/Simple-10                   1.42ms ± 0%    1.40ms ± 1%   -0.99%  (p=0.032 n=5+5)
EndToEnd/ored-preds-100/Prepared-10                  500µs ± 1%     473µs ± 1%   -5.44%  (p=0.008 n=5+5)
EndToEnd/ored-preds-using-params-100/Simple-10      2.46ms ± 1%    2.43ms ± 0%   -1.12%  (p=0.016 n=5+5)
EndToEnd/ored-preds-using-params-100/Prepared-10    1.33ms ± 1%    1.32ms ± 1%     ~     (p=0.222 n=5+5)

name                                              old alloc/op   new alloc/op   delta
EndToEnd/kv-read/Simple-10                          39.2kB ± 5%    39.2kB ± 4%     ~     (p=1.000 n=5+5)
EndToEnd/kv-read/Prepared-10                        27.4kB ± 3%    27.2kB ± 3%     ~     (p=0.635 n=5+5)
EndToEnd/kv-read-const/Simple-10                    27.7kB ± 4%    27.2kB ± 0%     ~     (p=0.413 n=5+4)
EndToEnd/kv-read-const/Prepared-10                  27.1kB ± 4%    27.0kB ± 3%     ~     (p=1.000 n=5+5)
EndToEnd/tpcc-new-order/Simple-10                   43.9kB ± 0%    43.9kB ± 0%     ~     (p=0.114 n=4+4)
EndToEnd/tpcc-new-order/Prepared-10                 28.2kB ± 0%    28.8kB ± 4%     ~     (p=0.190 n=4+5)
EndToEnd/tpcc-delivery/Simple-10                    96.5kB ±67%    61.7kB ± 0%     ~     (p=0.413 n=5+4)
EndToEnd/tpcc-delivery/Prepared-10                  47.2kB ± 0%    47.1kB ± 0%     ~     (p=0.905 n=5+4)
EndToEnd/tpcc-stock-level/Simple-10                  160kB ± 0%     159kB ± 0%   -0.56%  (p=0.029 n=4+4)
EndToEnd/tpcc-stock-level/Prepared-10                139kB ± 0%     139kB ± 0%   -0.56%  (p=0.016 n=4+5)
EndToEnd/many-columns-and-indexes-a/Simple-10       61.3kB ± 1%    61.1kB ± 1%     ~     (p=0.730 n=5+4)
EndToEnd/many-columns-and-indexes-a/Prepared-10     48.6kB ±11%    48.6kB ±10%     ~     (p=1.000 n=5+5)
EndToEnd/many-columns-and-indexes-b/Simple-10       64.2kB ± 0%    63.6kB ± 1%   -0.95%  (p=0.008 n=5+5)
EndToEnd/many-columns-and-indexes-b/Prepared-10     58.9kB ±35%    50.2kB ± 0%  -14.79%  (p=0.016 n=5+4)
EndToEnd/many-columns-and-indexes-c/Simple-10        111MB ±16%     111MB ±16%     ~     (p=0.841 n=5+5)
EndToEnd/many-columns-and-indexes-c/Prepared-10      106MB ± 0%     106MB ± 0%     ~     (p=1.000 n=5+5)
EndToEnd/single-col-histogram-range/Simple-10       46.9kB ± 8%    44.9kB ± 7%     ~     (p=0.151 n=5+5)
EndToEnd/single-col-histogram-range/Prepared-10     32.1kB ± 0%    32.1kB ± 0%     ~     (p=0.686 n=4+4)
EndToEnd/batch-insert-one/Simple-10                 7.25MB ± 4%    7.34MB ± 6%     ~     (p=1.000 n=5+5)
EndToEnd/batch-insert-one/Prepared-10               7.44MB ± 5%    7.43MB ± 5%     ~     (p=1.000 n=5+5)
EndToEnd/batch-insert-many/Simple-10                14.5MB ± 1%    15.4MB ±18%     ~     (p=0.730 n=4+5)
EndToEnd/batch-insert-many/Prepared-10              11.8MB ± 2%    13.6MB ±54%     ~     (p=1.000 n=4+5)
EndToEnd/ored-preds-100/Simple-10                    424kB ± 0%     401kB ± 3%   -5.56%  (p=0.016 n=4+5)
EndToEnd/ored-preds-100/Prepared-10                  254kB ± 0%     228kB ± 0%  -10.23%  (p=0.008 n=5+5)
EndToEnd/ored-preds-using-params-100/Simple-10      1.18MB ± 0%    1.15MB ± 0%   -2.38%  (p=0.029 n=4+4)
EndToEnd/ored-preds-using-params-100/Prepared-10    1.01MB ± 0%    1.00MB ± 4%     ~     (p=0.190 n=4+5)

name                                              old allocs/op  new allocs/op  delta
EndToEnd/kv-read/Simple-10                             414 ± 0%       417 ± 2%     ~     (p=1.000 n=4+5)
EndToEnd/kv-read/Prepared-10                           291 ± 0%       291 ± 1%     ~     (p=0.571 n=4+5)
EndToEnd/kv-read-const/Simple-10                       301 ± 0%       301 ± 0%     ~     (all equal)
EndToEnd/kv-read-const/Prepared-10                     284 ± 0%       284 ± 0%     ~     (p=1.000 n=4+4)
EndToEnd/tpcc-new-order/Simple-10                      455 ± 0%       454 ± 0%     ~     (p=0.486 n=4+4)
EndToEnd/tpcc-new-order/Prepared-10                    314 ± 0%       315 ± 0%     ~     (p=0.238 n=4+5)
EndToEnd/tpcc-delivery/Simple-10                       826 ±45%       627 ± 0%     ~     (p=0.476 n=5+4)
EndToEnd/tpcc-delivery/Prepared-10                     465 ± 0%       466 ± 0%     ~     (p=1.286 n=5+4)
EndToEnd/tpcc-stock-level/Simple-10                  1.23k ± 0%     1.22k ± 0%   -0.77%  (p=0.029 n=4+4)
EndToEnd/tpcc-stock-level/Prepared-10                1.04k ± 0%     1.03k ± 0%   -0.79%  (p=0.016 n=4+5)
EndToEnd/many-columns-and-indexes-a/Simple-10          542 ± 1%       540 ± 0%     ~     (p=0.159 n=5+4)
EndToEnd/many-columns-and-indexes-a/Prepared-10        418 ± 1%       417 ± 1%     ~     (p=0.400 n=4+4)
EndToEnd/many-columns-and-indexes-b/Simple-10          584 ± 0%       577 ± 0%   -1.20%  (p=0.016 n=5+4)
EndToEnd/many-columns-and-indexes-b/Prepared-10        457 ± 0%       451 ± 0%   -1.20%  (p=0.029 n=4+4)
EndToEnd/many-columns-and-indexes-c/Simple-10         718k ± 2%      718k ± 2%     ~     (p=0.571 n=4+4)
EndToEnd/many-columns-and-indexes-c/Prepared-10       712k ± 0%      712k ± 0%     ~     (p=0.690 n=5+5)
EndToEnd/single-col-histogram-range/Simple-10          447 ± 4%       444 ± 6%     ~     (p=0.389 n=5+5)
EndToEnd/single-col-histogram-range/Prepared-10        319 ± 0%       319 ± 0%     ~     (all equal)
EndToEnd/batch-insert-one/Simple-10                  44.4k ± 1%     44.1k ± 1%     ~     (p=0.190 n=5+4)
EndToEnd/batch-insert-one/Prepared-10                44.3k ± 0%     44.8k ± 6%     ~     (p=0.730 n=4+5)
EndToEnd/batch-insert-many/Simple-10                 77.0k ± 1%     80.2k ± 9%     ~     (p=1.000 n=4+5)
EndToEnd/batch-insert-many/Prepared-10               66.6k ± 4%     66.5k ± 5%     ~     (p=0.841 n=5+5)
EndToEnd/ored-preds-100/Simple-10                    4.25k ± 0%     3.85k ± 0%   -9.45%  (p=0.029 n=4+4)
EndToEnd/ored-preds-100/Prepared-10                  3.44k ± 0%     3.03k ± 0%  -11.67%  (p=0.008 n=5+5)
EndToEnd/ored-preds-using-params-100/Simple-10       7.50k ± 0%     7.09k ± 0%   -5.42%  (p=0.029 n=4+4)
EndToEnd/ored-preds-using-params-100/Prepared-10     6.55k ± 0%     6.27k ± 4%   -4.35%  (p=0.016 n=4+5)

name                                              old rows/op    new rows/op    delta
EndToEnd/tpcc-stock-level/Prepared-10                 1.00 ± 0%      1.00 ± 0%     ~     (all equal)
EndToEnd/batch-insert-one/Prepared-10                 1.00 ± 0%      1.00 ± 0%     ~     (all equal)
EndToEnd/batch-insert-many/Prepared-10                 181 ± 0%       181 ± 0%     ~     (all equal)
```

I think end-to-end benchmarks of the slow queries would show the bigger gains, but we currently don't have those benchmarks.